Create new review form and minimal Codepen facade
The Codepen facade assumes pure js, css, and html; no preprocessors.
Getting the preprocessors is tricky because there's no open API for it.
You can get the data from the `input#init-data` element on the pen page,
but that's not possible because of the page's CORS settings.

The new review page code is fixture for testing until there's some
proper code written.